+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/MageController.java
@@ -476,6 +476,9 @@ public interface MageController {
     boolean sendResourcePackToAllPlayers(CommandSender sender);
     boolean sendResourcePack(Player player);
     boolean promptResourcePack(Player player);
+    @Nullable
+    String getResourcePackURL();
+    boolean isResourcePackEnabledByDefault();
     boolean commitOnQuit();
 
     Set<String> getItemKeys();